Create Article research
Article 1.
This article was written to show the importance of “the importance of the manipulation-curiosity-exploration type of motive” (Thackray and Michels). The hypothesis of this article was Curiosity and manipulation can be used to motivate the learning of simple position responses. The control was the Procyon Lotor. The independent variable was curiosity and manipulation, and the dependent variable was learning of simple position response. The research does not contain external validity. This cannot be applied to other aspects besides the Procyon lotor. My hypothesis for this article was raccoons show a higher manipulation than rats. 
Thackray, Richard I., and Kenneth M. Michels. “Externally-Aroused Drives in the Racoon.” Animal Behaviour, vol. 6, no. 3-4, July 1958, pp. 160–163, 10.1016/0003-3472(58)90046-0. Accessed 17 Mar. 2022.

Article 2.
This research article was written to determine the number of food resources available to raccoons and if they are adaptable. The hypothesis of this article was “Abundance and distribution of food resources can profoundly affect movements and spatial distribution of solitary carnivores. The control of this experiment was food resources. The independent variable was food, and the dependent variable was the abundance and distribution of food. The only issue found with this article is there were no cofounds. It was more research-based with graphs. My hypothesis for this article is raccoons in urban cities have better access to food resources than in rural areas.
Prange, Suzanne, et al. “Influences of Anthropogenic Resources on Raccoon (Procyon Lotor) Movements and Spatial Distribution.” Journal of Mammalogy, vol. 83, no. 3, June 2004, academic.oup.com/jmammal/article/85/3/483/901013, 10.1644/1383946. Accessed 15 Nov. 2019.

Article 3.
           The article was written to provide more research into neuroscience research regarding brain function and cognitive flexibility. The hypothesis of this article was similarities with primate brains. The control was the raccoon. The independent variable was the cytoarchitectural characteristics and the dependent variable was adaptive behavioral outcomes. The only issue with this article was there was no randomization but the article did provide external validity. It can be used to understand depression and its function within the brain. Linking the two could help researchers understand more about each. My hypothesis for this article is Raccoons show better problem-solving skills than other animals.
Jacob, Joanna, et al. “Cytoarchitectural Characteristics Associated with Cognitive Flexibility in Raccoons.” Journal of Comparative Neurology, vol. 529, no. 14, 20 June 2021, pp. 3375–3388, 10.1002/cne.25197. Accessed 18 Nov. 2021.


Article 4.
This article was written about the somatic senses in the cerebral cortex of raccoons. The article brings insight into the behaviors of raccoons and creates a neurological approach. The hypothesis of this article was “An animal which has unusual or specialized anatomical structures and behavioral attributes would be expected to show a corresponding specialization in certain aspects of its neural structure and function” (Welker and Seidenstein). The control was the cerebral cortex. The independent variable was the nervous system, and the dependent variable was behavioral specialization. My hypothesis for this article is Raccoons have a more complex prefrontal cortex than humans. The article lacked randomization.
Welker, W. I., and Sidney Seidenstein. “Somatic Sensory Representation in the Cerebral Cortex of the Racoon (Procyon Lotor).” The Journal of Comparative Neurology, vol. 111, no. 3, June 1959, pp. 469–501, 10.1002/cne.901110306. Accessed 4 May 2020.


Article 5. 
This article is about the current and future climate that is favorable towards Procyon lotor. The hypothesis for this article was “the latter forecasts appear more relevant and should be privileged in the development of conservation plans for ecosystems” (Louppe et al.). The control in this article was the population of raccoons. The independent variable was climate favorability while the dependent variable was rural and urban cities. The article had internal validity, it had subject matching, exposure and outcomes within the study, and the control. My hypothesis is Raccoons can adapt better to climates regions than other animals.
Louppe, Vivien, et al. “Current and Future Climatic Regions Favourable for a Globally Introduced Wild Carnivore, the Raccoon Procyon Lotor.” Scientific Reports, vol. 9, no. 1, 24 June 2019, 10.1038/s41598-019-45713-y. Accessed 8 Oct. 2021.