Create Chose Hypothesis
The hypothesis chosen is Procyon lotor shows a higher level of manipulation than rats. The experimental design that can be used is the independent group design. Because we are measuring behavior in the raccoons this design would be more beneficial when proving the success rate of higher manipulation in raccoons. Block randomization cannot be used in this experiment due to it being behavior analysis based. The experimental design would include two boxes with a complex maze design with the outcome being food. Since both animals are more prone to go after food this experiment will be easier to run with an outcome. One for Procyon lotor and genius Rattus. They would have to figure out how to get through the maze with obstacles in the way such as harsh light to access the reward, which in this case is food. Since the T-maze is most common in research and experiments with rats, we will use that. This design would have observation with intervention because we are manipulating both animals. In the first trial, we will observe the animal’s ability to complete the maze with no obstacles or challenges. The second trial is when we intervene and add harsh lighting to confuse or make it difficult for the animals to complete the T-maze. We would not have an interrupted time series design since this is not time-based. Interrupted time series is more over a period whereas this experiment is just based on cognitive function and externally aroused drives.