Create Research Ethics
Research ethics to be considered in this experimental design include whether is there a threat to internal validity. Does the outcome match my hypothesis? Could there be bias with the animals already? Procyon is harder to find because they are more active at night and in the summer. As a researcher, did we find enough research to back up our claims? Is there evidence to support everything? Do we know enough to conduct research experiments on this animal? Plagiarism within this research if it was to be published could be avoided by crediting authors and sources where it is necessary and citing them at the end of the research. Will the experiment be done in a controlled environment?  Another thing to consider is, is our sources creditable and has it been peer-reviewed. As a researcher, we also must respect the animal’s dignity and still handle them with care. Did we as researchers make sure this animal is not an endangered species?  We must follow state, and federal laws, and regulations regarding animals and their safety. The culture and values of animals need to be considered. Although they are non-human, they still have a background and a way of life that we must respect and contribute to. The animals involved will not have to experience any pain, but slight discomfort with the lighting. There will be no types of sickness or death-related causes within this research.